pyproject.toml
[project]
name = "antispace-project"
version = "0.1.0"
description = "My first project"
authors = [
    {name = "SBrayanAlves"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests (>=2.32.4,<3.0.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "pydantic[email] (>=2.11.7,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
